Updates to internals and Websockets
Connection is best created in interactive environment by user. Creating it at the conclusion of the ws_create causes the polygon Websocket to error after a couple of seconds. Websocket will not connect from within a function, thus connecting it in ws_listen won't work either. This leaves it up to the user to create the connection. Examples updated accordingly. 
- logbars now properly heeded
- New internal for logging polygon data.